Document split NDISender API
