+++ Sources/NDI/NDISender.swift
@@ -3,229 +3,11 @@ import CoreVideo
 import Dependencies
 import libNDI
 
-public enum NDISendVideoFrameError: Error, Equatable, Sendable {
-	case unsupportedPixelFormat(OSType)
-	case invalidFrameRate
-	case couldNotLockPixelBuffer(CVReturn)
-	case missingPixelBufferBaseAddress
-}
-
-/// A video frame that keeps its pixel buffer alive until the synchronous NDI send completes.
-public struct NDISendVideoFrame: @unchecked Sendable {
-	private let pixelBuffer: CVPixelBuffer
-	private let fourCC: NDIlib_FourCC_video_type_e
-	private let frameRateNumerator: Int32
-	private let frameRateDenominator: Int32
-	private let timecode: NDITimecode
-
-	public init(
-		pixelBuffer: CVPixelBuffer,
-		frameRate: CMTime,
-		timecode: NDITimecode = .now
-	) throws {
-		guard
-			frameRate.isValid,
-			frameRate.value > 0,
-			frameRate.timescale > 0,
-			let numerator = Int32(exactly: frameRate.value)
-		else {
-			throw NDISendVideoFrameError.invalidFrameRate
-		}
-
-		let pixelFormat = CVPixelBufferGetPixelFormatType(pixelBuffer)
-		switch pixelFormat {
-		case kCVPixelFormatType_32BGRA:
-			fourCC = NDIlib_FourCC_video_type_BGRA
-		case kCVPixelFormatType_422YpCbCr8:
-			fourCC = NDIlib_FourCC_video_type_UYVY
-		default:
-			throw NDISendVideoFrameError.unsupportedPixelFormat(pixelFormat)
-		}
-
-		self.pixelBuffer = pixelBuffer
-		frameRateNumerator = numerator
-		frameRateDenominator = frameRate.timescale
-		self.timecode = timecode
-	}
-
-	fileprivate func withNDIFrame<R>(
-		_ operation: (UnsafePointer<NDIlib_video_frame_v2_t>) -> R
-	) throws -> R {
-		let lockResult = CVPixelBufferLockBaseAddress(pixelBuffer, .readOnly)
-		guard lockResult == kCVReturnSuccess else {
-			throw NDISendVideoFrameError.couldNotLockPixelBuffer(lockResult)
-		}
-		defer {
-			CVPixelBufferUnlockBaseAddress(pixelBuffer, .readOnly)
-		}
-
-		guard let baseAddress = CVPixelBufferGetBaseAddress(pixelBuffer) else {
-			throw NDISendVideoFrameError.missingPixelBufferBaseAddress
-		}
-
-		var frame = NDIlib_video_frame_v2_t(
-			xres: Int32(CVPixelBufferGetWidth(pixelBuffer)),
-			yres: Int32(CVPixelBufferGetHeight(pixelBuffer)),
-			FourCC: fourCC,
-			frame_rate_N: frameRateNumerator,
-			frame_rate_D: frameRateDenominator,
-			picture_aspect_ratio: Float(CVPixelBufferGetWidth(pixelBuffer)) / Float(CVPixelBufferGetHeight(pixelBuffer)),
-			frame_format_type: NDIlib_frame_format_type_progressive,
-			timecode: timecode.rawValue,
-			p_data: baseAddress.assumingMemoryBound(to: UInt8.self),
-			NDIlib_video_frame_v2_t.__Unnamed_union___Anonymous_field9(
-				line_stride_in_bytes: Int32(CVPixelBufferGetBytesPerRow(pixelBuffer))
-			),
-			p_metadata: nil,
-			timestamp: 0
-		)
-		return withUnsafePointer(to: &frame, operation)
-	}
-}
-
-public enum NDISendAudioFrameError: Error, Equatable, Sendable {
-	case invalidSampleRate
-	case noChannels
-	case inconsistentSampleCounts
-	case tooManyChannels
-	case tooManySamples
-}
-
-/// Planar, 32-bit floating-point audio that remains alive until a synchronous NDI send completes.
-public struct NDISendAudioFrame: Sendable {
-	public let sampleRate: Int
-	public let numberOfChannels: Int
-	public let numberOfSamples: Int
-	public let timecode: NDITimecode
-	public let metadata: String?
-
-	private let samples: [Float]
-
-	public init(
-		planarSamples: [[Float]],
-		sampleRate: Int = 48000,
-		timecode: NDITimecode = .now,
-		metadata: String? = nil
-	) throws {
-		guard sampleRate > 0, Int32(exactly: sampleRate) != nil else {
-			throw NDISendAudioFrameError.invalidSampleRate
-		}
-		guard let firstChannel = planarSamples.first else {
-			throw NDISendAudioFrameError.noChannels
-		}
-		guard Int32(exactly: planarSamples.count) != nil else {
-			throw NDISendAudioFrameError.tooManyChannels
-		}
-		guard Int32(exactly: firstChannel.count) != nil else {
-			throw NDISendAudioFrameError.tooManySamples
-		}
-		guard planarSamples.allSatisfy({ $0.count == firstChannel.count }) else {
-			throw NDISendAudioFrameError.inconsistentSampleCounts
-		}
-
-		self.sampleRate = sampleRate
-		numberOfChannels = planarSamples.count
-		numberOfSamples = firstChannel.count
-		self.timecode = timecode
-		self.metadata = metadata
-		samples = planarSamples.flatMap { $0 }
-	}
-
-	fileprivate func withNDIFrame<R>(_ operation: (UnsafePointer<NDIlib_audio_frame_v3_t>) -> R) -> R {
-		samples.withUnsafeBufferPointer { samples in
-			withMetadataCString { metadata in
-				var frame = NDIlib_audio_frame_v3_t(
-					sample_rate: Int32(sampleRate),
-					no_channels: Int32(numberOfChannels),
-					no_samples: Int32(numberOfSamples),
-					timecode: timecode.rawValue,
-					FourCC: NDIlib_FourCC_audio_type_FLTP,
-					p_data: samples.baseAddress.map {
-						UnsafeMutableRawPointer(mutating: $0).assumingMemoryBound(to: UInt8.self)
-					},
-					.init(channel_stride_in_bytes: Int32(numberOfSamples * MemoryLayout<Float>.stride)),
-					p_metadata: metadata,
-					timestamp: 0
-				)
-				return withUnsafePointer(to: &frame, operation)
-			}
-		}
-	}
-
-	private func withMetadataCString<R>(_ operation: (UnsafePointer<CChar>?) -> R) -> R {
-		if let metadata {
-			return metadata.withCString(operation)
-		}
-		return operation(nil)
-	}
-}
-
-public struct NDISendMetadataFrame: Hashable, Sendable {
-	public var value: String
-	public var timecode: NDITimecode
-
-	public init(value: String, timecode: NDITimecode = .now) {
-		self.value = value
-		self.timecode = timecode
-	}
-
-	fileprivate func withNDIFrame<R>(_ operation: (UnsafePointer<NDIlib_metadata_frame_t>) -> R) -> R {
-		value.utf8CString.withUnsafeBufferPointer { value in
-			var frame = NDIlib_metadata_frame_t(
-				length: Int32(value.count),
-				timecode: timecode.rawValue,
-				p_data: value.baseAddress.map(UnsafeMutablePointer.init(mutating:))
-			)
-			return withUnsafePointer(to: &frame, operation)
-		}
-	}
-}
-
-public struct NDISenderTally: Equatable, Sendable {
-	public var isOnProgram: Bool
-	public var isOnPreview: Bool
-
-	public init(isOnProgram: Bool, isOnPreview: Bool) {
-		self.isOnProgram = isOnProgram
-		self.isOnPreview = isOnPreview
-	}
-}
-
-public enum NDISenderCaptureResult: Sendable {
-	case none
-	case metadata(NDISenderCapturedMetadataFrame)
-	case statusChange
-	case error
-	case unknown
-}
-
-public final class NDISenderCapturedMetadataFrame: @unchecked Sendable {
-	private var ref: NDIlib_metadata_frame_t
-	private let sender: NDISender
-
-	fileprivate init(ref: NDIlib_metadata_frame_t, sender: NDISender) {
-		self.ref = ref
-		self.sender = sender
-	}
-
-	deinit {
-		sender.freeMetadata(&ref)
-	}
-
-	public var timecode: NDITimecode {
-		NDITimecode(rawValue: ref.timecode)
-	}
-
-	public var value: String? {
-		guard let data = ref.p_data else { return nil }
-		if ref.length == 0 {
-			return String(cString: data)
-		}
-		let length = max(0, Int(ref.length) - 1)
-		return String(bytes: Data(bytes: data, count: length), encoding: .utf8)
-	}
-}
-
+/// An NDI source that sends video, audio, and metadata to receivers on the network.
+///
+/// The sender does not clock audio or video submissions. Pace calls to `send(_:)`
+/// at the rate represented by each frame. The sender remains advertised until it
+/// is deallocated.
 public final class NDISender: @unchecked Sendable {
 	// NDIlib_send_instance_t is generally thread safe as long as it's not freed before NDIlib_send_send_video_v2_async finishes  (https://docs.ndi.video/all/developing-with-ndi/sdk/ndi-send).
 
@@ -234,8 +16,9 @@ public final class NDISender: @unchecked Sendable {
 
 	/// Creates a sender using the process-wide NDI runtime.
 	///
-	/// Returns `nil` when the NDI runtime cannot be loaded or the SDK cannot create
-	/// a sender with the supplied name.
+	/// - Parameter name: The user-visible source name advertised on the NDI network.
+	/// - Returns: A sender, or `nil` when the NDI runtime cannot be loaded or the
+	///   SDK cannot create a sender with the supplied name.
 	public convenience init?(name: String) {
 		@Dependency(\.ndi) var ndi
 		guard let ndi else { return nil }
@@ -262,6 +45,15 @@ public final class NDISender: @unchecked Sendable {
 		ndi.NDIlib_send_destroy(sender)
 	}
 
+	/// Returns the current number of receivers connected to this source.
+	///
+	/// Use the result to avoid rendering or generating media while there are no
+	/// receivers. When `timeout` is greater than zero and no receiver is connected,
+	/// the call waits up to that duration for a connection.
+	///
+	/// - Parameter timeout: How long to wait for a receiver connection. Pass zero
+	///   to poll without waiting.
+	/// - Returns: The number of connected receivers.
 	public func connectionCount(timeout: Duration = .zero) -> Int {
 		Int(ndi.NDIlib_send_get_no_connections(
 			sender,
@@ -269,6 +61,12 @@ public final class NDISender: @unchecked Sendable {
 		))
 	}
 
+	/// Returns the source name advertised by this sender.
+	///
+	/// NDI normally qualifies the configured name with the machine name when it
+	/// advertises the source.
+	///
+	/// - Returns: The advertised source name, or `nil` if the SDK does not provide one.
 	public func sourceName() -> String? {
 		guard let source = ndi.NDIlib_send_get_source_name(sender) else {
 			return nil
@@ -277,24 +75,38 @@ public final class NDISender: @unchecked Sendable {
 		return String(cString: source.pointee.p_ndi_name)
 	}
 
-	/// Adds a video frame synchronously.
+	/// Sends a video frame synchronously.
+	///
+	/// The call returns after NDI finishes accessing the frame's pixel buffer.
+	///
+	/// - Parameter frame: The video frame to send to connected receivers.
+	/// - Throws: ``NDISendVideoFrameError`` if the pixel buffer cannot be locked
+	///   or accessed for the send.
 	public func send(_ frame: NDISendVideoFrame) throws {
 		try frame.withNDIFrame { frame in
 			ndi.NDIlib_send_send_video_v2(sender, frame)
 		}
 	}
 
-	/// Adds a planar floating-point audio frame synchronously.
+	/// Sends a planar floating-point audio frame synchronously.
+	///
+	/// - Parameter frame: The audio frame to send to connected receivers.
 	public func send(_ frame: NDISendAudioFrame) {
 		frame.withNDIFrame { ndi.NDIlib_send_send_audio_v3(sender, $0) }
 	}
 
 	/// Sends metadata to all connected receivers.
+	///
+	/// - Parameter frame: A UTF-8 XML metadata frame.
 	public func send(_ frame: NDISendMetadataFrame) {
 		frame.withNDIFrame { ndi.NDIlib_send_send_metadata(sender, $0) }
 	}
 
-	/// Receives metadata sent back by a connected receiver.
+	/// Waits for a message sent upstream by a connected receiver.
+	///
+	/// - Parameter timeout: How long to wait for a message. Pass zero to poll
+	///   without waiting.
+	/// - Returns: The captured message or the reason no metadata was returned.
 	public func capture(timeout: Duration = .zero) -> NDISenderCaptureResult {
 		var metadata = NDIlib_metadata_frame_t(length: 0, timecode: 0, p_data: nil)
 		let frameType = ndi.NDIlib_send_capture(
@@ -317,7 +129,12 @@ public final class NDISender: @unchecked Sendable {
 		}
 	}
 
-	/// Returns a tally update, or `nil` if the state did not change before the timeout.
+	/// Waits for the program or preview tally state to change.
+	///
+	/// - Parameter timeout: How long to wait for a change. Pass zero to poll the
+	///   current tally state without waiting.
+	/// - Returns: The current tally after a change, or `nil` when no change occurs
+	///   before the timeout.
 	public func tally(timeout: Duration = .zero) -> NDISenderTally? {
 		var tally = NDIlib_tally_t(on_program: false, on_preview: false)
 		guard ndi.NDIlib_send_get_tally(sender, &tally, UInt32(timeout.seconds * 1000)) else {
@@ -326,17 +143,30 @@ public final class NDISender: @unchecked Sendable {
 		return NDISenderTally(isOnProgram: tally.on_program, isOnPreview: tally.on_preview)
 	}
 
-	/// Clears all metadata automatically sent to newly connected receivers.
+	/// Clears all connection metadata queued for receivers.
+	///
+	/// Connection metadata is sent automatically whenever a new receiver connects.
 	public func clearConnectionMetadata() {
 		ndi.NDIlib_send_clear_connection_metadata(sender)
 	}
 
-	/// Adds metadata that is sent to existing and future receiver connections.
+	/// Adds metadata to send automatically with receiver connections.
+	///
+	/// The metadata is queued for each new connection and is also sent immediately
+	/// to receivers that are already connected. Call ``clearConnectionMetadata()``
+	/// to reset the queue.
+	///
+	/// - Parameter frame: A UTF-8 XML metadata frame describing the connection.
 	public func addConnectionMetadata(_ frame: NDISendMetadataFrame) {
 		frame.withNDIFrame { ndi.NDIlib_send_add_connection_metadata(sender, $0) }
 	}
 
-	/// Sets the source receivers should use if this sender becomes unavailable.
+	/// Sets the source that receivers should use if this sender becomes unavailable.
+	///
+	/// Receivers can switch back automatically if this sender returns. Pass `nil`
+	/// to clear the failover source.
+	///
+	/// - Parameter source: The fallback NDI source, or `nil` to remove the fallback.
 	public func setFailoverSource(_ source: NDISource?) {
 		guard var source = source?.ref else {
 			ndi.NDIlib_send_set_failover(sender, nil)
@@ -345,25 +175,7 @@ public final class NDISender: @unchecked Sendable {
 		ndi.NDIlib_send_set_failover(sender, &source)
 	}
 
-	fileprivate func freeMetadata(_ metadata: UnsafePointer<NDIlib_metadata_frame_t>) {
+	func freeMetadata(_ metadata: UnsafePointer<NDIlib_metadata_frame_t>) {
 		ndi.NDIlib_send_free_metadata(sender, metadata)
-	}
-}
-
-private final class NDISenderConnectionSubscriptionToken: @unchecked Sendable {
-	let id = UUID()
-	private let lock = NSLock()
-	private var cancelled = false
-
-	func cancel() {
-		lock.lock()
-		cancelled = true
-		lock.unlock()
-	}
-
-	var isCancelled: Bool {
-		lock.lock()
-		defer { lock.unlock() }
-		return cancelled
 	}
 }
